// Gerdau/web/src/redux/machine/types.ts
export enum MachineActionTypes {
  FETCH_MACHINES = "FETCH_MACHINES",
  FETCH_MACHINES_SUCCESS = "FETCH_MACHINES_SUCCESS",
  FETCH_MACHINES_ERROR = "FETCH_MACHINES_ERROR",

  FETCH_MACHINE = "FETCH_MACHINE",
  FETCH_MACHINE_SUCCESS = "FETCH_MACHINE_SUCCESS",
  FETCH_MACHINE_ERROR = "FETCH_MACHINE_ERROR",

  FETCH_MODELS = "FETCH_MODELS",
  FETCH_MODELS_SUCCESS = "FETCH_MODELS_SUCCESS",
  FETCH_MODELS_ERROR = "FETCH_MODELS_ERROR",

  FETCH_MAKES = "FETCH_MAKES",
  FETCH_MAKES_SUCCESS = "FETCH_MAKES_SUCCESS",
  FETCH_MAKES_ERROR = "FETCH_MAKES_ERROR",
}